pyproject.toml
[tool.poetry]
name = "botato"
version = "0.1.0"
description = "A modern and maintainable Discord bot framework for Python."
authors = ["Shiv <shiv@zerobytes.fr>"]
license = "GPL-3.0"
readme = "README.md"
packages = [{ include = "botato" }]
repository = "https://github.com/shivkun/botato"
keywords = ["discord", "bot", "framework", "python", "asyncio"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Framework :: AsyncIO",
    "Topic :: Software Development :: Libraries",
    "Intended Audience :: Developers",
    "Development Status :: 3 - Alpha"
]

[tool.poetry.dependencies]
python = "^3.13"
aiohttp = "^3.11.18"
websockets = "^15.0"
pydantic = "^2.11"
aiolimiter = "^1.2"
loguru = "^0.7.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
mypy = "^1.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"